feat: 收口小尾巴

+++ BackpackSurvivor/Assets/BackpackSurvivor/Scripts/GamePlay/Waves/EnemySpawner.cs
@@ -38,7 +38,7 @@ namespace BS.GamePlay.Waves
             }
         }
 
-        public void ApplyWaveSettings(float spawnInterval , float maxAlive)
+        public void ApplyWaveSettings(float spawnInterval , int maxAlive)
         {
             if(spawnInterval < 0.1||maxAlive <=0) return;
             this.spawnInterval = spawnInterval;
